+++ docs/html/search/all_2.js
@@ -7,9 +7,11 @@ var searchData=
   ['common_4',['common',['../group__timer__to__ms.html',1,'Timer Aux common'],['../group__timer__common.html',1,'Timer common']]],
   ['counter0_20mode0_5',['Counter0 Mode0',['../group__counter0__mode0.html',1,'']]],
   ['counter0_5fmode0_2eh_6',['counter0_mode0.h',['../counter0__mode0_8h.html',1,'']]],
-  ['counter0_5fmode0_5fget_5fvalue_7',['counter0_mode0_get_value',['../group__counter0__mode0.html#gaaab5c77175ed786722627dc5a77e1933',1,'counter0_mode0.h']]],
-  ['counter0_5fmode0_5finit_8',['counter0_mode0_init',['../group__counter0__mode0.html#ga91c9a1b9a89def2cb319b1eca4177de5',1,'counter0_mode0.h']]],
-  ['counter0_5fmode0_5fset_5fvalue_9',['counter0_mode0_set_value',['../group__counter0__mode0.html#gaacabd9f5df806591acaf4d072e6abf91',1,'counter0_mode0.h']]],
-  ['counter0_5fmode0_5fstart_10',['counter0_mode0_start',['../group__counter0__mode0.html#ga6303fc5dfc8f4b0dafd341940e35f417',1,'counter0_mode0.h']]],
-  ['counter0_5fmode0_5fstop_11',['counter0_mode0_stop',['../group__counter0__mode0.html#ga0529c8b6076e792659c29d8b1f39c9b0',1,'counter0_mode0.h']]]
+  ['counter0_5fmode0_5fdisable_5fp35_5foutput_7',['counter0_mode0_disable_P35_output',['../group__counter0__mode0.html#ga28e73dbb30c38e09d4611427813e6175',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5fenable_5fp35_5foutput_8',['counter0_mode0_enable_P35_output',['../group__counter0__mode0.html#gae405fc1390dc5469617d2f862d77a33e',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5fget_5fvalue_9',['counter0_mode0_get_value',['../group__counter0__mode0.html#gaaab5c77175ed786722627dc5a77e1933',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5finit_10',['counter0_mode0_init',['../group__counter0__mode0.html#ga91c9a1b9a89def2cb319b1eca4177de5',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5fset_5fvalue_11',['counter0_mode0_set_value',['../group__counter0__mode0.html#gaacabd9f5df806591acaf4d072e6abf91',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5fstart_12',['counter0_mode0_start',['../group__counter0__mode0.html#ga6303fc5dfc8f4b0dafd341940e35f417',1,'counter0_mode0.h']]],
+  ['counter0_5fmode0_5fstop_13',['counter0_mode0_stop',['../group__counter0__mode0.html#ga0529c8b6076e792659c29d8b1f39c9b0',1,'counter0_mode0.h']]]
 ];